requirements.txt
Flask
git+https://github.com/mikicz/arca.git#egg=arca
pyyaml
pytest
pytest-cov
pytest-flake8
requests
